refactor: remove payment methods page and related components
- Deleted the payment methods page component and its associated HTML, SCSS, and TypeScript files.
- Updated routing to remove references to the payment methods page.
- Refactored checkout subscription component to navigate to the new payment methods page.
- Introduced a new payment methods list component with improved structure and functionality.
- Added constants for default avatar URLs.
- Updated detail creator profile component to use the new default avatar constant.
- Cleaned up unused imports and console logs in various components.
