--- apps/athlete/src/app/features/profile/pages/payment-methods-page/payment-methods-page.component.html
@@ -1,97 +1,0 @@
-<ion-content class="payment-methods-content">
-  <app-layout-content
-    customClass="payment-methods-page"
-    [useGoBack]="true"
-    [defaultHref]="
-      '/home/suscriptions/suscription-creator/' + _sesisonService.user$().userId
-    "
-    [title]="'go-back' | translate"
-  >
-    <ion-grid content-body>
-      <ion-grid>
-        <ion-row>
-          <ion-col size="12">
-            <ion-button
-              color="tertiary"
-              class="bb-input"
-              expand="block"
-              shape="round"
-              (click)="openModalAddPaymentMethod()"
-            >
-              {{ 'add-payment-method' | translate }}
-            </ion-button>
-          </ion-col>
-          <ion-col size="12">
-            <ion-radio-group
-              [ngModel]="selectedPaymentMethod()"
-              (ngModelChange)="selectedPaymentMethod.set($event)"
-            >
-              @for (method of paymentMethods(); track method.id) {
-              <ion-item
-                lines="none"
-                (click)="selectPaymentMethod(method)"
-                [ngClass]="{
-                  selected: method.id === selectedPaymentMethod()
-                }"
-                class="payment-method-item"
-              >
-                <ion-icon
-                  name="card-outline"
-                  slot="start"
-                  size="large"
-                ></ion-icon>
-
-                <ion-label>
-                  <h2>
-                    <strong>{{ method.brand | titlecase }}</strong>
-                    @if (method.is_default) {
-                    <ion-badge color="success" class="default-badge">
-                      {{ 'default' | translate }}
-                    </ion-badge>
-                    }
-                  </h2>
-                  <h3>•••• •••• •••• {{ method.last4 }}</h3>
-                  <p>
-                    {{ 'expires' | translate }} {{ method.exp_month }}/{{
-                      method.exp_year
-                    }}
-                  </p>
-                </ion-label>
-
-                <ion-radio
-                  slot="end"
-                  class="bb-radio"
-                  [value]="method.id"
-                ></ion-radio>
-              </ion-item>
-              } @if (paymentMethods().length === 0) {
-              <ion-item lines="none" class="empty-state-item">
-                <ion-icon
-                  name="card-outline"
-                  slot="start"
-                  size="large"
-                  color="medium"
-                ></ion-icon>
-                <ion-label>
-                  <h2>{{ 'no-payment-methods' | translate }}</h2>
-                  <p>{{ 'add-first-payment-method' | translate }}</p>
-                </ion-label>
-              </ion-item>
-              }
-            </ion-radio-group>
-          </ion-col>
-        </ion-row>
-      </ion-grid>
-    </ion-grid>
-    <div content-footer class="ion-text-center">
-      <ion-button
-        color="tertiary"
-        expand="block"
-        [disabled]="!selectedPaymentMethod()"
-        [routerLink]="['/home/suscriptions/checkout']"
-      >
-        <ion-text class="question">{{ 'Suscribete' | translate }}</ion-text>
-      </ion-button>
-    </div>
-  </app-layout-content>
-</ion-content>
