+++ apps/athlete/src/app/features/suscriptions/page/checkout-suscription/checkout-suscription.component.html
@@ -21,7 +21,7 @@
               <div class="trainer-info">
                 <ion-avatar class="trainer-avatar">
                   <img
-                    [src]="trainer().profilePictureUrl"
+                    [src]="trainer().profilePictureUrl || defaultProfilePicture"
                     [alt]="trainer().firstName + ' ' + trainer().lastName"
                   />
                 </ion-avatar>
@@ -63,11 +63,7 @@
               </ion-card-title>
             </ion-card-header>
             <ion-card-content>
-              <ion-item
-                (click)="selectPaymentMethod()"
-                lines="none"
-                class="selected-payment-method"
-              >
+              <ion-item lines="none" class="selected-payment-method">
                 <ion-icon
                   name="card-outline"
                   slot="start"
@@ -85,11 +81,6 @@
                     }}
                   </p>
                 </ion-label>
-                <ion-icon
-                  name="chevron-forward"
-                  slot="end"
-                  color="medium"
-                ></ion-icon>
               </ion-item>
               <ion-button
                 fill="clear"
