+++ apps/athlete/src/app/features/suscriptions/page/detail-suscription-creator/detail-suscription-creator.component.html
@@ -12,16 +12,20 @@
     text=""
     style="position: absolute; top: 16px; left: 16px; z-index: 10"
   ></ion-back-button>
-  <img
-    src="assets/images/suscription/suscription.png"
-    alt=""
-    style="
-      width: 100%;
-      height: 337px;
-      object-fit: cover;
-      object-position: center;
-    "
-  />
+  <div class="container-header-subscription">
+    <img src="assets/images/logo/logo.svg" alt="" class="img-logo" />
+    <ion-text class="subscription-title"
+      >Suscripción<br /><ion-text class="creator-name" color="secondary">{{
+        fullName()
+      }}</ion-text></ion-text
+    >
+
+    <ion-text class="description" color="white"
+      >Obtén acceso a la selección de entrenamientos y clases que se ofrecen en
+      la aplicación.</ion-text
+    >
+  </div>
+
   @if( creator() ) {
   <app-layout-content customClass="suscription-creator-page ion-padding">
     <ion-grid content-body>
@@ -66,7 +70,7 @@
         class="bb-input"
         expand="block"
         shape="round"
-        routerLink="/home/profile/payment-methods"
+        routerLink="/home/suscriptions/payment-methods"
       >
         {{ 'Suscribirse' | translate }}
       </ion-button>
