+++ apps/athlete/src/app/features/suscriptions/page/detail-suscription-creator/detail-suscription-creator.component.scss
@@ -5,9 +5,50 @@
     ::ng-deep {
       .suscription-creator-page {
         min-height: auto;
+        margin-bottom: 45px;
       }
     }
+    .container-header-subscription {
+      background: linear-gradient(180deg, rgba(0, 0, 0, 0) 0%, #000 101.65%),
+        url('/assets/images/suscription/suscription.png');
+      background-size: cover;
+      background-position: center;
+      height: 337px;
 
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      color: white;
+      padding: 24px;
+      gap: 16px;
+      .img-logo {
+        height: 68px;
+        width: 60px;
+        margin-bottom: 16px;
+      }
+
+      .subscription-title {
+        font-size: 16px;
+        font-style: normal;
+        font-weight: 700;
+        line-height: 20px;
+        letter-spacing: 0.2px;
+      }
+      .creator-name {
+        font-size: 24px;
+        font-style: normal;
+        font-weight: 700;
+        line-height: 28px;
+      }
+
+      .description {
+        font-size: 16px;
+        font-style: normal;
+        font-weight: 400;
+        line-height: 20px;
+        letter-spacing: 0.2px;
+      }
+    }
     ion-item {
       margin-bottom: 24px;
       border-radius: 12px;
