+++ libs/ui/src/lib/components/layout-content/layout-content.component.scss
@@ -1,7 +1,7 @@
 .layout-content {
   padding-top: var(--ion-safe-area-top);
   padding-bottom: var(--ion-safe-area-bottom);
-  margin-top: 15px;
+  // margin-top: 15px;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
