+++ libs/ui/src/lib/components/payment-methods-list/payment-methods-list.component.html
@@ -1,1 +1,70 @@
-<p>payment-methods-list works!</p>,
+<ion-grid>
+  <ion-row>
+    <ion-col size="12">
+      <ion-button
+        color="tertiary"
+        class="bb-input"
+        expand="block"
+        shape="round"
+        (click)="openModalAddPaymentMethod()"
+      >
+        {{ 'add-payment-method' | translate }}
+      </ion-button>
+    </ion-col>
+    <ion-col size="12">
+      <ion-radio-group
+        [ngModel]="selectedPaymentMethod()"
+        (ngModelChange)="selectedPaymentMethod.set($event)"
+      >
+        @for (method of paymentMethods(); track method.id) {
+        <ion-item
+          lines="none"
+          (click)="selectPaymentMethod(method)"
+          [ngClass]="{
+            selected: method.id === selectedPaymentMethod()
+          }"
+          class="payment-method-item"
+        >
+          <ion-icon name="card-outline" slot="start" size="large"></ion-icon>
+
+          <ion-label>
+            <h2>
+              <strong>{{ method.brand | titlecase }}</strong>
+              @if (method.is_default) {
+              <ion-badge color="success" class="default-badge">
+                {{ 'default' | translate }}
+              </ion-badge>
+              }
+            </h2>
+            <h3>•••• •••• •••• {{ method.last4 }}</h3>
+            <p>
+              {{ 'expires' | translate }} {{ method.exp_month }}/{{
+                method.exp_year
+              }}
+            </p>
+          </ion-label>
+
+          <ion-radio
+            slot="end"
+            class="bb-radio"
+            [value]="method.id"
+          ></ion-radio>
+        </ion-item>
+        } @if (paymentMethods().length === 0) {
+        <ion-item lines="none" class="empty-state-item">
+          <ion-icon
+            name="card-outline"
+            slot="start"
+            size="large"
+            color="medium"
+          ></ion-icon>
+          <ion-label>
+            <h2>{{ 'no-payment-methods' | translate }}</h2>
+            <p>{{ 'add-first-payment-method' | translate }}</p>
+          </ion-label>
+        </ion-item>
+        }
+      </ion-radio-group>
+    </ion-col>
+  </ion-row>
+</ion-grid>
