+++ libs/ui/src/lib/components/payment-methods-list/payment-methods-list.component.scss
@@ -1,3 +1,75 @@
 :host {
   display: block;
 }
+.payment-method-item {
+  margin-bottom: 12px;
+  border-radius: 12px;
+  --background: var(--ion-color-light, #f8f9fa);
+  --border-style: none;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
+
+  &[color='primary'] {
+    --background: var(--ion-color-primary, #3880ff);
+    --color: white;
+
+    ion-label h2,
+    ion-label h3,
+    ion-label p {
+      color: white;
+    }
+  }
+
+  ion-label {
+    h2 {
+      font-weight: 600;
+      margin-bottom: 4px;
+      display: flex;
+      align-items: center;
+      gap: 8px;
+    }
+
+    h3 {
+      font-family: 'Courier New', monospace;
+      font-size: 16px;
+      font-weight: 500;
+      margin-bottom: 4px;
+      color: var(--ion-color-dark, #222428);
+    }
+
+    p {
+      font-size: 12px;
+      color: var(--ion-color-medium, #92949c);
+      margin: 0;
+    }
+  }
+}
+
+.default-badge {
+  font-size: 10px;
+  --background: var(--ion-color-success, #2dd36f);
+  --color: white;
+}
+
+.empty-state-item {
+  text-align: center;
+  padding: 40px 20px;
+  --background: transparent;
+
+  ion-label {
+    h2 {
+      color: var(--ion-color-medium, #92949c);
+      font-weight: 600;
+      margin-bottom: 8px;
+    }
+
+    p {
+      color: var(--ion-color-medium, #92949c);
+      font-size: 14px;
+    }
+  }
+}
+
+ion-item.selected::part(native) {
+  border: 3px solid var(--ion-color-secondary);
+  border-radius: 12px;
+}
